requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
SQLAlchemy>=2.0
pyodbc>=4.0
python-dotenv>=1.0
python-docx==1.1.2
PyMuPDF==1.24.8
spire.doc==12.1.0
boto3
openpyxl==3.1.2
docxcompose==1.4.0
